#5 section 4-5 - rename handler adapter files
